231226 박주영: 로딩페이지 기초 세팅
