+++ src/Page/Loading/Components/Web/WebLoading.js
@@ -1,14 +1,53 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
+import BackgroundImg from "../../../../Assets/Loading_Background"
 
 const WebLoading = () => {
-    return (
 
-        <div>
-            
-        </div>
-    )
-
-}
+  return (
+    <Div>
+      <Img src={BackgroundImg}></Img>
+      <Div></Div>
+    </Div>
+  );
+};
 
 export default WebLoading;
+
+// theme 파일 폰트 적용 방법 + style-components 사용
+const Header1 = styled.div`
+  font-size: ${({ theme }) => theme.Web_fontSizes.Header1};
+  font-weight: ${({ theme }) => theme.fontWeights.Header1};
+  line-height: ${({ theme }) => theme.LineHeight.Header1};
+  color: ${({ theme }) => theme.colors.secondary};
+  font-family: "Pretendard";
+`;
+
+const Div = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  margin: 0vh 0vh 0vh 0vh;
+  padding: 0vh 0vh 0vh 0vh;
+  background-color: white;
+  /* border: 0.5px solid black; */
+  border-radius: 0px;
+  box-sizing: content-box;
+  font-size: ${({ theme }) => theme.Web_fontSizes.Header1};
+  font-weight: ${({ theme }) => theme.fontWeights.Header1};
+  line-height: ${({ theme }) => theme.LineHeight.Header1};
+  font-family: Pretendard;
+  /* color: ${({ theme }) => theme.colors.secondary}; */
+  grid-template-rows: 1fr;
+  grid-template-columns: 1fr 1fr;
+`;
+
+const A = styled.a``;
+
+const Img = styled.img`
+  width: 100%;
+  height: 100%;
+`;
